Add extensibility for case presentation formatting
Introduced ICasePresentationFormatter and its default implementation to allow custom formatting of answer text and HTML cards in the case workflow. Refactored CasePresenterStep to use the new formatter and registered it in DI. Added localized, parameterized error messages for ownership verification failures after max attempts. Cleaned up obsolete formatting code and improved localization handling.
